requirements.txt
africastalking==1.2.8
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
idna==3.8
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
sqlparse==0.5.1
urllib3==2.2.2
